finsight_backend/requirements/base.txt
# Core Django packages
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.2
django-knox==4.2.0

# Database
psycopg2-binary==2.9.9

# Authentication & Security
bcrypt==4.1.3
cryptography==42.0.8

# API Documentation
drf-spectacular==0.27.2

# Task Queue
celery==5.3.4
redis==5.0.7

# Image handling
Pillow==10.4.0

# Environment management
python-decouple==3.8

# Utilities
python-dateutil==2.9.0
pytz==2024.1
